requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
biopython==1.83
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
entrezpy==2.1.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
lxml==5.2.1
numpy==1.26.4
openai==1.17.1
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
scipy==1.13.0
sniffio==1.3.1
soupsieve==2.5
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.6.0
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
